Add workqueue threading
Before, each thread was assigned a cell on an mxn pixel grid. This led to
unbalanced performance for pictures with regions of empty space. Now, threading
is done by having worker threads pull the next pixel available for rendering.

This commit also paves the way for writing partial updates to the image (either
OpenGL or png) at higher and higher sampling levels.